Fixed star of certain points not being generated correctly

+++ src/engine/graphics/shapes/Star.java
@@ -12,9 +12,9 @@ public class Star
             Coordinate coordinate1 = new Coordinate( offset.getX(), offset.getY() );
             Coordinate coordinate2 = new Coordinate( offset.getX(), offset.getY() );
             Coordinate coordinate3 = new Coordinate( offset.getX(), offset.getY() );
-            coordinate1.moveByAngle( 360 / points * i, highPoints );
-            coordinate2.moveByAngle( (int)( 360 / points * (i + 0.5f) ), lowPoints );
-            coordinate3.moveByAngle( 360 / points * ( i + 1 ), highPoints );
+            coordinate1.moveByAngle( (int)( 360 / (float)points * i ), highPoints );
+            coordinate2.moveByAngle( (int)( 360 / (float)points * (i + 0.5f) ), lowPoints );
+            coordinate3.moveByAngle( (int)( 360 / (float)points * ( i + 1 ) ), highPoints );
 
             Line.draw( coordinate1, coordinate2, color, renderer );
             Line.draw( coordinate2, coordinate3, color, renderer );
